requirements.txt
gitpython
transformers
torch==2.6.0
torchvision==0.21.0
rich
tqdm
huggingface_hub
requests
fastapi
uvicorn[standard]